// src/middlewares/queries.generated.ts
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** country code */
  CountryCode: any;
  /** An RFC-3339 compliant Full Date Scalar */
  Date: any;
  /** An RFC-3339 compliant DateTime Scalar */
  DateTime: any;
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: any;
  /** gender */
  Gender: any;
  /** Long type */
  Long: any;
};
// ...
export type MiddlewareRedirectOrRewriteQueryVariables = Types.Exact<{
  host: Types.Scalars['String'];
  pathname: Types.Scalars['String'];
}>;


export type MiddlewareRedirectOrRewriteQuery = { redirect?: { newUrl: string, status: number } | null, page?: { __typename: 'BPD_ListPage' } | { __typename: 'BPD_NonPage' } | { __typename: 'BPD_ProductPage' } | { __typename: 'BPD_RegionPage' } | { __typename: 'BPD_StaticPage' } | null };

export type MiddlewareChannelByHostQueryVariables = Types.Exact<{
  host: Types.Scalars['String'];
}>;


export type MiddlewareChannelByHostQuery = { channel: { id: string, theme: string, key: string, url: string } };


export const MiddlewareRedirectOrRewriteDocument = gql`
    query MiddlewareRedirectOrRewrite($host: String!, $pathname: String!) {
  redirect: bpd_redirect(pathname: $pathname) {
    newUrl
    status
  }
  page: bpd_page(host: $host, pathname: $pathname) {
    __typename
  }
}
    `;